tests/wpt/web-platform-tests/css/css-shadow-parts/host-stylesheet.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS Shadow Parts - Host Stylesheet</title>
    <meta href="mailto:fergal@chromium.org" rel="author" title="Fergal Daly">
    <link href="http://www.google.com/" rel="author" title="Google">
    <link href="https://drafts.csswg.org/css-shadow-parts/" rel="help">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="support/shadow-helper.js"></script>
  </head>
  <body>
    <script>installCustomElement("custom-element", "custom-element-template");</script>
    <template id="custom-element-template">
      <style>
        ::part(partp) { color: red; }
        span { color: green; }
      </style>
      <span id="part" part="partp">This text</span>
    </template>
    The following text should be green:
    <custom-element id="c-e"></custom-element>
    <script type="text/javascript">
      "use strict";
      const colorGreen = "rgb(0, 128, 0)";
      test(function() {
        var el = getElementByShadowIds(document, ["c-e", "part"]);
        assert_equals(window.getComputedStyle(el).color, colorGreen);
      }, "Part in selected host is not styled by ::part in a stylesheet inside the host");
    </script>
  </body>
</html>
